requirements.txt
# Core AI
anthropic>=0.39.0

# Web search
requests>=2.31.0
beautifulsoup4>=4.12.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# Environment and config
python-dotenv>=1.0.0

# Rich terminal output
rich>=13.7.0

# Telegram bot
python-telegram-bot>=21.0

# Pipeline
pyyaml>=6.0
